# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Agent Assembly API
  description: |-
    REST API for the Agent Assembly governance gateway.

    This spec is auto-generated from `aa-api` route annotations via `utoipa`. CI fails if the generated spec drifts from the committed `openapi/v1.yaml`.
  contact:
    name: Agent Assembly Contributors
    url: https://github.com/AI-agent-assembly/agent-assembly
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
- url: http://localhost:7700
  description: Local development gateway
paths:
  /api/v1/agents:
    get:
      tags:
      - agents
      summary: '`GET /api/v1/agents` — list all registered agents with pagination.'
      description: Returns a paginated list of all agents currently known to the registry.
      operationId: list_agents
      parameters:
      - name: page
        in: query
        description: Page number (1-indexed). Defaults to 1.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 1
      - name: per_page
        in: query
        description: Items per page (max 100). Defaults to 50.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: Paginated list of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentResponse'
...
  /api/v1/iam/api-keys:
    get:
      tags:
      - iam
      summary: '`GET /api/v1/iam/api-keys` — list every API key the IAM store knows.'
      description: |-
        Returned in newest-first order. Active and revoked keys are both
        included; the dashboard filters by status in its tabs.
      operationId: list_api_keys
      responses:
        '200':
          description: All API keys, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyResponse'
    post:
      tags:
      - iam
      summary: '`POST /api/v1/iam/api-keys` — issue a new API key.'
      description: |-
        IAM mutations are gated as a `Global`-scope policy update — the caller
        must hold the `OrgAdmin` role.

        The response `secret` field is shown to the caller **once**; the server
        does not persist it.
      operationId: generate_api_key
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateApiKeyRequest'
        required: true
      responses:
        '200':
          description: Generated key (with one-shot secret)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApiKeyResponse'
        '403':
          description: Caller lacks the role required to mutate IAM state
  /api/v1/iam/api-keys/{id}/revoke:
    post:
      tags:
      - iam
      summary: '`POST /api/v1/iam/api-keys/{id}/revoke` — revoke an existing key.'
      description: 404 if the key is unknown, 409 if it is already revoked.
      operationId: revoke_api_key
      parameters:
      - name: id
        in: path
        description: API key id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Key revoked
        '403':
          description: Caller lacks the role required to mutate IAM state
        '404':
          description: Unknown api key id
        '409':
          description: Key is already revoked
  /api/v1/iam/api-keys/{id}/rotate:
    post:
      tags:
      - iam
      summary: |-
        `POST /api/v1/iam/api-keys/{id}/rotate` — atomically revoke `id` and
        issue a replacement carrying the same label, scopes, and owner.
      description: |-
        Returns the new key's one-shot reveal. 404 if `id` is unknown, 409 if
        the source key is already revoked.
      operationId: rotate_api_key
      parameters:
      - name: id
        in: path
        description: API key id to rotate
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Replacement key with one-shot secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApiKeyResponse'
        '403':
          description: Caller lacks the role required to mutate IAM state
        '404':
          description: Unknown api key id
        '409':
          description: Source key is already revoked
...
  /api/v1/policies:
    get:
      tags:
      - policies
      summary: '`GET /api/v1/policies` — list all policy versions.'
      description: List all governance policy versions with pagination.
      operationId: list_policies
      parameters:
      - name: page
        in: query
        description: Page number (1-indexed). Defaults to 1.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 1
      - name: per_page
        in: query
        description: Items per page (max 100). Defaults to 50.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: Paginated list of policy versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyResponse'
    post:
      tags:
      - policies
      summary: '`POST /api/v1/policies` — apply a new governance policy.'
      description: |-
        Submit and activate a new governance policy from YAML.
        The caller must hold the role required for the target `scope`
        (default: `global`, requires `OrgAdmin`).
      operationId: create_policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
        required: true
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          description: Invalid policy YAML or scope
        '403':
          description: Insufficient role for this policy scope
...
components:
  schemas:
    ActiveSessionResponse:
      type: object
      description: Summary of an active session in the API response.
      required:
      - session_id
      - started_at
      - status
      properties:
        session_id:
          type: string
          description: Hex-encoded session UUID.
        started_at:
          type: string
          description: ISO 8601 timestamp when the session started.
        status:
          type: string
          description: Current status of the session.
...
    AgentResponse:
      type: object
      description: JSON representation of an agent returned by the API.
      required:
      - id
      - name
      - framework
      - version
      - status
      - tool_names
      - metadata
      - session_count
      - policy_violations_count
      - active_sessions
      - recent_events
      - recent_traces
      properties:
        active_sessions:
          type: array
          items:
            $ref: '#/components/schemas/ActiveSessionResponse'
          description: Currently active sessions for this agent.
        framework:
          type: string
          description: Agent framework (e.g. "langgraph", "crewai").
        id:
          type: string
          description: Hex-encoded agent UUID.
        last_event:
          type:
          - string
          - 'null'
          description: ISO 8601 timestamp of the most recent event.
        layer:
          type:
          - string
          - 'null'
          description: Governance layer this agent is assigned to (e.g. "advisory", "enforced").
        metadata:
          type: object
          description: Arbitrary metadata key-value pairs.
          additionalProperties:
            type: string
          propertyNames:
            type: string
        name:
          type: string
          description: Human-readable agent name.
        pid:
          type:
          - integer
          - 'null'
          format: int32
          description: OS process ID, if known.
          minimum: 0
        policy_violations_count:
          type: integer
          format: int32
          description: Number of policy violations recorded.
          minimum: 0
        recent_events:
          type: array
          items:
            $ref: '#/components/schemas/RecentEventResponse'
          description: Most recent events emitted by this agent.
        recent_traces:
          type: array
          items:
            $ref: '#/components/schemas/RecentTraceResponse'
          description: Most recent trace session IDs for this agent.
        session_count:
          type: integer
          format: int32
          description: Number of sessions handled.
          minimum: 0
        status:
          type: string
          description: Current runtime status.
        tool_names:
          type: array
          items:
            type: string
          description: Tools declared at registration.
        version:
          type: string
          description: Semver version string.
...
    ApiKeyResponse:
      type: object
      required:
      - id
      - label
      - prefix
      - scopes
      - status
      - created_at
      - owner
      - role
      - assigned_policies
      - recent_activity
      properties:
        assigned_policies:
          type: array
          items:
            type: string
        created_at:
          type: string
        id:
          type: string
        label:
          type: string
        last_used:
          type:
          - string
          - 'null'
        owner:
          type: string
        prefix:
          type: string
        recent_activity:
          type: array
          items:
            $ref: '#/components/schemas/RecentActivityResponse'
        role:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyScopeResponse'
        status:
          $ref: '#/components/schemas/ApiKeyStatusResponse'
    ApiKeyScopeResponse:
      type: string
      description: Scopes a key may hold. Wire form matches the dashboard's TS union.
      enum:
      - read:members
      - write:members
      - read:policies
      - write:policies
      - read:audit
      - admin
    ApiKeyStatusResponse:
      type: string
      enum:
      - active
      - revoked
...
    CreatePolicyRequest:
      type: object
      description: Request body for creating a new policy.
      required:
      - policy_yaml
      properties:
        policy_yaml:
          type: string
          description: Raw YAML content of the governance policy.
        scope:
          type:
          - string
          - 'null'
          description: |-
            Governance scope this policy targets (e.g. `"global"`, `"team:platform"`).

            Used for RBAC authorization — the caller must hold the role required
            to mutate policies at this scope. Defaults to `"global"` when absent.
...
    GenerateApiKeyRequest:
      type: object
      required:
      - label
      - scopes
      properties:
        label:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyScopeResponse'
    GeneratedApiKeyResponse:
      type: object
      description: |-
        One-shot reveal shape returned by generate / rotate. `secret` MUST be
        captured by the caller — the server does not store it.
      required:
      - id
      - prefix
      - secret
      properties:
        id:
          type: string
        prefix:
          type: string
        secret:
          type: string
...
    PolicyResponse:
      type: object
      description: JSON representation of a governance policy version.
      required:
      - name
      - version
      - active
      - rule_count
      - policy_yaml
      properties:
        active:
          type: boolean
          description: Whether this is the currently active policy.
        name:
          type: string
          description: Policy name from metadata.
        policy_yaml:
          type: string
          description: |-
            Raw YAML content of this policy version. Empty string when the
            underlying snapshot is not retrievable from the history store
            (e.g. a policy loaded at startup before any history entry exists).
        rule_count:
          type: integer
          description: Number of rules in this policy version.
          minimum: 0
        version:
          type: string
          description: Policy version string.
...
    RecentActivityResponse:
      type: object
      required:
      - id
      - timestamp
      - action
      - target
      properties:
        action:
          type: string
        id:
          type: string
        target:
          type: string
        timestamp:
          type: string
    RecentEventResponse:
      type: object
      description: Summary of a recent event in the API response.
      required:
      - event_type
      - summary
      - timestamp
      properties:
        event_type:
          type: string
          description: Event type classification (e.g. "violation", "approval", "budget").
        summary:
          type: string
          description: Short human-readable summary.
        timestamp:
          type: string
          description: ISO 8601 timestamp when the event occurred.
    RecentTraceResponse:
      type: object
      description: Summary of a recent trace session for an agent.
      required:
      - session_id
      - timestamp
      properties:
        session_id:
          type: string
          description: Hex-encoded session UUID, usable with `aasm trace <session-id>`.
        timestamp:
          type: string
          description: ISO 8601 timestamp when the trace session started.
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key (`aa_…` prefix) or JWT bearer token
...